+++ changelog.md
@@ -8,6 +8,9 @@
   variables. They're injected into the zone running the current task.
 - `copyFile` and `copyDirectory` and deprecated in favor of a new `copy` method
 - `deleteEntity` is deprecated in favor of a new `delete` method
+- Renamed `runProcess`/`runProcessAsync`/`runDartScript` to `run`/`runAsync`/`Dart.run`.
+  Process result info (stdout, stderr, exitCode) is now exposed by these
+  methods and some others which call them.
 - Added a wrapper class around `pub global activate` applications - `PubApplication`
 - Grinder can now create a simple starter script for a project - run `pub run grinder:init`